Set back amount from double to int  due to format error from vnpay

+++ src/main/java/com/example/thanhhoa/services/vnpay/VNPayService.java
@@ -4,7 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 
 public interface VNPayService {
 
-    String createOrder(Double total, String reason, String urlReturn, String ipAddr);
+    String createOrder(int total, String reason, String urlReturn, String ipAddr);
 
     int orderReturn(HttpServletRequest request);
 }
